(Refactor): Restructure test directories to integration, unit and utils folders [pv622]
