+++ aignostic/tests/unit/model_endpoint/test_pydantic_conversions.py
@@ -1,12 +1,14 @@
 import aignostic.pydantic_models.data_models as data_models
 import numpy as np
-from tests.model_api.model.mock import app as mock_app
+from tests.utils.model.mock import app as mock_app
 from fastapi.testclient import TestClient
 import pandas as pd
+import pytest
 
 mock_app = TestClient(mock_app)
 
 
+@pytest.mark.skip(reason="Functionality no longer required/supported - test pending deletion review")
 def test_conversion_from_numpy():
     """
     Test conversion from numpy array to pydantic model
@@ -18,6 +20,7 @@ def test_conversion_from_numpy():
     assert response.status_code == 200, response.text
 
 
+@pytest.mark.skip(reason="Functionality no longer required/supported - test pending deletion review")
 def test_conversion_from_pandas():
     """
     Test conversion from pandas dataframe to pydantic model
@@ -29,6 +32,7 @@ def test_conversion_from_pandas():
     assert response.status_code == 200, response.text
 
 
+@pytest.mark.skip(reason="Functionality no longer required/supported - test pending deletion review")
 def test_conversion_from_csv():
     """
     Test conversion from csv to pydantic model
